fix different signs: 'int' and 'unsigned long'

+++ src/mvsd/mgbubble/MgServer.cpp
@@ -233,9 +233,9 @@ void MgServer::ev_notify_handler(struct mg_connection *nc, int ev, void *ev_data
         if (nc->flags & MG_F_USER_2)
         {
             assert(nc->recv_mbuf.len >= sizeof(struct mg_event));
-            assert(*((int*)(ev_data)) >= sizeof(struct mg_event));
-            int len = 0;
-            int maxlen = nc->recv_mbuf.len / sizeof(struct mg_event) * sizeof(struct mg_event);
+            assert(*((size_t*)(ev_data)) >= sizeof(struct mg_event));
+            size_t len = 0;
+            size_t maxlen = nc->recv_mbuf.len / sizeof(struct mg_event) * sizeof(struct mg_event);
             for (; len < maxlen; len += sizeof(struct mg_event)) {
                 struct mg_event* pev = (struct mg_event*)(nc->recv_mbuf.buf + len);
                 self->on_notify_handler(*nc, *pev);
